Handle incoming messages from bot
